% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeserror.R
\name{average_mutual_information}
\alias{average_mutual_information}
\title{Average Mutual Information}
\usage{
average_mutual_information(X, ave)
}
\description{
The average mutual information as a fraction of the total entropy.
}
