% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeserror.R
\name{bayeserror}
\alias{bayeserror}
\title{Estimate the Bayes error for a data set using a list of classifiers.}
\usage{
bayeserror(
  x,
  y,
  classifiers,
  err_method = "accuracy",
  corr_method = "ami",
  v = 10L
)
}
\arguments{
\item{x}{a matrix of features}

\item{y}{a vector of the observed classes}

\item{classifiers}{a list of classifier functions or else their names}

\item{err_method}{classification loss method}

\item{corr_method}{correlation of error method}

\item{v}{the resampling method}
}
\description{
Estimate the Bayes error for a data set using a list of classifiers.
}
