% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeserror.R
\name{dual_total_correlation}
\alias{dual_total_correlation}
\title{Dual Total Correlation for Binary Data}
\usage{
dual_total_correlation(X, normalized = FALSE)
}
\arguments{
\item{X}{a matrix whose columns are 0/1 class vectors}
}
\description{
Dual Total Correlation for Binary Data
}
