% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeserror.R
\name{ensemble_errors}
\alias{ensemble_errors}
\title{Compute classifier error from a matrix of class predictions.}
\usage{
ensemble_errors(predictions, true_classes, method = "accuracy")
}
\description{
Compute classifier error from a matrix of class predictions.
}
