% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeserror.R
\name{error_correlation}
\alias{error_correlation}
\title{Get mutual-information based correlated-error for predicted classes.}
\usage{
error_correlation(total, ave, method = "ami")
}
\arguments{
\item{total}{matrix of base classifier 0/1 predictions}

\item{ave}{vector of ensemble 0/1 predictions}

\item{method}{the method to use to compute the correlation}
}
\description{
Get mutual-information based correlated-error for predicted classes.
}
