% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeserror.R
\name{mahalanobis_bound}
\alias{mahalanobis_bound}
\title{Upper Bound of the Bayes Error Rate given by the Mahalanobis Distance}
\usage{
mahalanobis_bound(x, y)
}
\arguments{
\item{x}{a data matrix with numeric values}

\item{y}{a binary vector}
}
\value{
an estimate of the Bayes error rate of x w.r.t. y
}
\description{
Provides an upper bound for the Bayes error rate of a dataset on
binary classes. This bound requires no assumptions of the
distribution of the data.
}
\details{
The Mahalanobis distance between class 0 and class 1 is given by
\deqn{M_d = (\mu_0 - \mu_1)^T \Sigma^{-1}(\mu_0 - \mu_1)}{%
M_d = (mu_0 - mu_1)' Sigma^-1 (mu_0 - mu_1)}
where \eqn{\mu_0} and \eqn{\mu_1} are the class means in \code{x} for
class 0 and 1, respectively. This provides an upper bound on the Bayes
error \eqn{\epsilon_bayes} by
\deqn{\epsilon_bayes \leq \frac{2 Pr(0) Pr(1)}{1 + Pr(0) Pr(1) M_d}}{%
\epsilon_bayes <= (2 Pr(0) Pr(1)) / (1 + Pr(0) Pr(1) M_d)}
where \eqn{Pr(0)} and \eqn{Pr(1)} are the class probabilities of 0 and
1.
}
