% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeserror.R
\name{merge_folds}
\alias{merge_folds}
\title{Merge folded data}
\usage{
merge_folds(folded_data)
}
\arguments{
\item{folded_data}{a list of arrays containing the results of a
computation on folded data, one array for each fold.}
}
\description{
Merge folded data
}
\keyword{internal}
