% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeserror.R
\name{predict_ensemble}
\alias{predict_ensemble}
\title{Generate class prediction probabilities}
\usage{
predict_ensemble(x, y, classifiers, v = 10L)
}
\arguments{
\item{x}{A matrix with observations as rows and features as
columns.}

\item{y}{A vector of 0/1 response classes}

\item{classifiers}{A list of classifier functions. Each function
should be of the form \code{f(x, y, newdata)} and have the attribute
\code{shortname} to identify its predictions.}

\item{v}{For k-fold cross-validation, use an integer to specify the
number of folds. For instance, \code{v = 10} means to use 10-fold
CV. For a holdout split, use a number between 0.0 to 1.0
specifying the percent to be held out to make predictions. For
in-sample predictions, use \code{v = 0}.}
}
\value{
A three-dimensional array containing class prediction
probabilities. Its dimesions are as (observation, class,
classifier). The last entry on the third dimension is \code{"AVE"},
the predictions for the ensemble classifier.
}
\description{
This function will generate class prediction probabilities on a
dataset from a list of classifiers using a specified resampling
method.
}
