% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayeserror.R
\name{predict_on_fold}
\alias{predict_on_fold}
\title{Generate predictions on a cross-validation fold.}
\usage{
predict_on_fold(x, y, fold, classifiers)
}
\value{
an array of class probability predictions
}
\description{
Generate predictions on a cross-validation fold.
}
\keyword{internal}
